pyproject.toml
[project]
name = "search-appearance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.4.0",
    "httpx>=0.28.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "uvicorn[standard]>=0.40.0",
    "sentence-transformers>=2.2.2",
    "pillow>=10.0.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.0",
    "websockets>=12.0",
    "pyyaml>=6.0.0",
    "sqlmodel>=0.0.31",
    "lancedb>=0.27.1",
    "alembic>=1.18.1",
    "aiohttp>=3.9.0",
    "facenet-pytorch>=2.5.3",
]

[tool.setuptools]
packages = ["mvp"]

